refactor: add protocol service
- [x] Add protocol service for proof generation
- [x] Cleanup injector service
