+++ packages/app/src/background/services/injector/__tests__/injector.test.ts
@@ -2,22 +2,9 @@
  * @jest-environment jsdom
  */
 
-import {
-  PendingRequestType,
-  ConnectedIdentityMetadata,
-  IRLNProofRequest,
-  ISemaphoreProofRequest,
-  IZkMetadata,
-  IJoinGroupMemberArgs,
-  IGenerateGroupMerkleProofArgs,
-} from "@cryptkeeperzk/types";
-import { getMerkleProof } from "@cryptkeeperzk/zk";
-import { omit } from "lodash";
-import browser from "webextension-polyfill";
+import { PendingRequestType, type ConnectedIdentityMetadata, type IZkMetadata } from "@cryptkeeperzk/types";
 
-import { createChromeOffscreen, getBrowserPlatform } from "@src/background/shared/utils";
 import { mockDefaultIdentity } from "@src/config/mock/zk";
-import { RPCInternalAction, BrowserPlatform } from "@src/constants";
 import pushMessage from "@src/util/pushMessage";
 
 import { InjectorService } from "..";
@@ -40,8 +27,6 @@ const mockNewRequest = jest.fn((type: PendingRequestType, { urlOrigin }: { urlOr
 });
 
 const mockGetConnectedIdentity = jest.fn();
-const mockGenerateSemaphoreProof = jest.fn();
-const mockGenerateRLNProof = jest.fn();
 const mockGetConnectedIdentityData = jest.fn(
   (_: unknown, meta?: IZkMetadata): Promise<ConnectedIdentityMetadata | undefined> => {
     if (meta?.urlOrigin === mockDefaultUrlOrigin || meta?.urlOrigin === "new-urlOrigin") {
@@ -67,9 +52,6 @@ const mockAwaitLockServiceUnlock = jest.fn();
 const mockAwaitZkIdentityServiceUnlock = jest.fn();
 const mockAwaitApprovalServiceUnlock = jest.fn();
 
-const mockJoinGroupRequest = jest.fn();
-const mockGenerateGroupProofRequest = jest.fn();
-
 Object.defineProperty(global, "chrome", {
   value: {
     offscreen: {
@@ -83,14 +65,6 @@ Object.defineProperty(global, "chrome", {
   },
 });
 
-jest.mock("@cryptkeeperzk/zk", (): unknown => ({
-  ZkProofService: jest.fn(() => ({
-    generateSemaphoreProof: mockGenerateSemaphoreProof,
-    generateRLNProof: mockGenerateRLNProof,
-  })),
-  getMerkleProof: jest.fn(),
-}));
-
 jest.mock("@src/background/controllers/browserUtils", (): unknown => ({
   getInstance: jest.fn(() => ({
     closePopup: jest.fn(),
@@ -101,13 +75,6 @@ jest.mock("@src/background/controllers/browserUtils", (): unknown => ({
 jest.mock("@src/background/controllers/requestManager", (): unknown => ({
   getInstance: jest.fn(() => ({
     newRequest: mockNewRequest,
-  })),
-}));
-
-jest.mock("@src/background/services/group", (): unknown => ({
-  getInstance: jest.fn(() => ({
-    joinGroupRequest: mockJoinGroupRequest,
-    generateGroupMerkleProofRequest: mockGenerateGroupProofRequest,
   })),
 }));
 
@@ -151,7 +118,6 @@ describe("background/services/injector", () => {
 
   beforeEach(() => {
     (pushMessage as jest.Mock).mockReset();
-    (getMerkleProof as jest.Mock).mockResolvedValue({});
 
     mockGetConnectedIdentity.mockResolvedValue({ serialize: () => mockSerializedIdentity });
   });
@@ -183,7 +149,7 @@ describe("background/services/injector", () => {
       );
     });
 
-    test("should send undefined if urlOrigin isn't approved", async () => {
+    test("should send undefined if origin isn't approved", async () => {
       mockGetStatus.mockResolvedValueOnce({ isUnlocked: false, isInitialized: true });
 
       const service = InjectorService.getInstance();
@@ -282,445 +248,6 @@ describe("background/services/injector", () => {
 
       await expect(service.connect({ isChangeIdentity: false }, { urlOrigin: "reject_connect" })).rejects.toThrow(
         "CryptKeeper: error in the connect request, User rejected your request.",
-      );
-    });
-  });
-
-  describe("generate Semaphore proof", () => {
-    beforeEach(() => {
-      mockGetStatus.mockResolvedValueOnce({ isUnlocked: true, isInitialized: true });
-      (browser.runtime.getURL as jest.Mock).mockImplementation((path: string) => path);
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    const defaultProofRequest: ISemaphoreProofRequest = {
-      identitySerialized: "identitySerialized",
-      externalNullifier: "externalNullifier",
-      signal: "signal",
-      merkleProofSource: "https://merklestorageurl.com",
-      circuitFilePath: "js/zkeyFiles/semaphore/semaphore.wasm",
-      zkeyFilePath: "js/zkeyFiles/semaphore/semaphore.zkey",
-      verificationKey: "js/zkeyFiles/semaphore/semaphore.json",
-    };
-
-    const emptyFullProof = {
-      fullProof: {
-        proof: {},
-        publicSignals: {},
-      },
-    };
-
-    test("should throw error if there is no origin url in metadata", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, {})).rejects.toThrowError("Origin is not set");
-    });
-
-    test("should throw error if there is no connected identity", async () => {
-      mockGetConnectedIdentity.mockResolvedValue(undefined);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: connected identity is not found",
-      );
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error if urlOrigin isn't approved", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, { urlOrigin: "new-urlOrigin" })).rejects.toThrow(
-        "CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.",
-      );
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error there is no circuit and zkey files", async () => {
-      mockGetConnectedIdentity.mockResolvedValue({
-        serialize: () => mockSerializedIdentity,
-        genIdentityCommitment: () => "mockIdentityCommitment",
-      });
-      (browser.runtime.getURL as jest.Mock).mockReturnValue(undefined);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: Must set Semaphore circuitFilePath and zkeyFilePath",
-      );
-    });
-
-    test("should throw error if user rejected semaphore approve request", async () => {
-      mockGetConnectedIdentity.mockResolvedValue({
-        serialize: () => mockSerializedIdentity,
-        genIdentityCommitment: () => "mockIdentityCommitment",
-      });
-      const service = InjectorService.getInstance();
-
-      await expect(
-        service.generateSemaphoreProof(defaultProofRequest, { urlOrigin: "reject_semaphore_proof" }),
-      ).rejects.toThrow("CryptKeeper: error in the Semaphore approve request User rejected your request.");
-    });
-
-    test("should generate semaphore proof properly on Firefox platform browsers", async () => {
-      mockGetConnectedIdentity.mockResolvedValue({
-        serialize: () => mockSerializedIdentity,
-        genIdentityCommitment: () => "mockIdentityCommitment",
-      });
-      mockGenerateSemaphoreProof.mockReturnValueOnce(emptyFullProof);
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Firefox);
-      const service = InjectorService.getInstance();
-
-      const result = await service.generateSemaphoreProof(defaultProofRequest, defaultMetadata);
-      expect(result).toStrictEqual(emptyFullProof);
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error if generate proof is failed on Firefox", async () => {
-      const error = new Error("error");
-      (pushMessage as jest.Mock).mockReturnValueOnce(JSON.stringify(emptyFullProof));
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Firefox);
-      mockGenerateSemaphoreProof.mockRejectedValue(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating Semaphore proof on Firefox error`,
-      );
-    });
-
-    test("should generate semaphore proof properly on Chrome platform browsers", async () => {
-      (pushMessage as jest.Mock).mockReturnValueOnce(emptyFullProof);
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Chrome);
-      const service = InjectorService.getInstance();
-
-      const result = await service.generateSemaphoreProof(defaultProofRequest, defaultMetadata);
-
-      expect(result).toStrictEqual(emptyFullProof);
-      expect(pushMessage).toBeCalledTimes(1);
-      expect(pushMessage).toBeCalledWith({
-        method: RPCInternalAction.GENERATE_SEMAPHORE_PROOF_OFFSCREEN,
-        payload: {
-          ...omit(defaultProofRequest, ["merkleProofSource"]),
-          urlOrigin: mockDefaultUrlOrigin,
-          identitySerialized: mockSerializedIdentity,
-          merkleStorageUrl: "https://merklestorageurl.com",
-          merkleProofArtifacts: undefined,
-          merkleProofProvided: undefined,
-        },
-        meta: mockDefaultUrlOrigin,
-        source: "offscreen",
-      });
-    });
-
-    test("should throw error if generate proof is failed on Chrome platform browsers", async () => {
-      const error = new Error("error");
-
-      (pushMessage as jest.Mock).mockRejectedValueOnce(error);
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Chrome);
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating Semaphore proof on Chrome error`,
-      );
-      expect(pushMessage).toBeCalledTimes(1);
-      expect(pushMessage).toBeCalledWith({
-        method: RPCInternalAction.GENERATE_SEMAPHORE_PROOF_OFFSCREEN,
-        payload: {
-          ...omit(defaultProofRequest, ["merkleProofSource"]),
-          urlOrigin: mockDefaultUrlOrigin,
-          identitySerialized: mockSerializedIdentity,
-          merkleStorageUrl: "https://merklestorageurl.com",
-          merkleProofArtifacts: undefined,
-          merkleProofProvided: undefined,
-        },
-        meta: mockDefaultUrlOrigin,
-        source: "offscreen",
-      });
-    });
-
-    test("should throw error if generate proof is failed to create Chrome offscreen in Chrome platform browsers", async () => {
-      const error = new Error("error");
-      (pushMessage as jest.Mock).mockReturnValueOnce(emptyFullProof);
-      (createChromeOffscreen as jest.Mock).mockRejectedValueOnce(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateSemaphoreProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating Semaphore proof on Chrome error`,
-      );
-
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-  });
-
-  describe("generate RLN Proof", () => {
-    beforeEach(() => {
-      mockGetStatus.mockResolvedValueOnce({ isUnlocked: true, isInitialized: true });
-      (browser.runtime.getURL as jest.Mock).mockImplementation((path: string) => path);
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    const defaultProofRequest: IRLNProofRequest = {
-      identitySerialized: "identitySerialized",
-      rlnIdentifier: "rlnIdentifier",
-      message: "message",
-      messageId: 0,
-      messageLimit: 1,
-      merkleProofSource: "https://merklestorageurl.com",
-      circuitFilePath: "js/zkeyFiles/rln/rln.wasm",
-      zkeyFilePath: "js/zkeyFiles/rln/rln.zkey",
-      verificationKey: "js/zkeyFiles/rln/rln.json",
-      epoch: Date.now().toString(),
-    };
-
-    const emptyFullProof = {
-      proof: {},
-      publicSignals: {},
-    };
-
-    test("should generate rln proof properly on Chrome platform browsers", async () => {
-      (pushMessage as jest.Mock).mockReturnValueOnce(JSON.stringify(emptyFullProof));
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Chrome);
-      const service = InjectorService.getInstance();
-
-      const result = await service.generateRLNProof(defaultProofRequest, defaultMetadata);
-
-      expect(result).toStrictEqual(emptyFullProof);
-      expect(pushMessage).toBeCalledTimes(1);
-      expect(pushMessage).toBeCalledWith({
-        method: RPCInternalAction.GENERATE_RLN_PROOF_OFFSCREEN,
-        payload: {
-          ...omit(defaultProofRequest, ["merkleProofSource"]),
-          urlOrigin: mockDefaultUrlOrigin,
-          identitySerialized: mockSerializedIdentity,
-          merkleStorageUrl: "https://merklestorageurl.com",
-          merkleProofArtifacts: undefined,
-          merkleProofProvided: undefined,
-        },
-        meta: mockDefaultUrlOrigin,
-        source: "offscreen",
-      });
-    });
-
-    test("should throw error if generate RLN proof is failed on Chrome platform browsers", async () => {
-      const error = new Error("error");
-
-      (pushMessage as jest.Mock).mockRejectedValueOnce(error);
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Chrome);
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating RLN proof on Chrome error`,
-      );
-      expect(pushMessage).toBeCalledTimes(1);
-      expect(pushMessage).toBeCalledWith({
-        method: RPCInternalAction.GENERATE_RLN_PROOF_OFFSCREEN,
-        payload: {
-          ...omit(defaultProofRequest, ["merkleProofSource"]),
-          urlOrigin: mockDefaultUrlOrigin,
-          identitySerialized: mockSerializedIdentity,
-          merkleStorageUrl: "https://merklestorageurl.com",
-          merkleProofArtifacts: undefined,
-          merkleProofProvided: undefined,
-        },
-        meta: mockDefaultUrlOrigin,
-        source: "offscreen",
-      });
-    });
-
-    test("should throw error if generate proof is failed to create Chrome offscreen in Chrome platform browsers", async () => {
-      const error = new Error("error");
-      (pushMessage as jest.Mock).mockReturnValueOnce(emptyFullProof);
-      (createChromeOffscreen as jest.Mock).mockRejectedValueOnce(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating RLN proof on Chrome error`,
-      );
-
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error if there is no origin url in metadata", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, {})).rejects.toThrowError("Origin is not set");
-    });
-
-    test("should throw error if there is no connected identity", async () => {
-      mockGetConnectedIdentity.mockResolvedValue(undefined);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: connected identity is not found",
-      );
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error if user rejected semaphore approve request", async () => {
-      mockGetConnectedIdentity.mockResolvedValue({
-        serialize: () => mockSerializedIdentity,
-        genIdentityCommitment: () => "mockIdentityCommitment",
-      });
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, { urlOrigin: "reject_rln_proof" })).rejects.toThrow(
-        "CryptKeeper: error in the RLN approve request User rejected your request.",
-      );
-    });
-
-    test("should throw error if urlOrigin isn't approved", async () => {
-      (pushMessage as jest.Mock).mockReturnValueOnce(JSON.stringify(emptyFullProof));
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, { urlOrigin: "new-urlOrigin" })).rejects.toThrow(
-        "CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.",
-      );
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should generate proof for firefox platform properly", async () => {
-      (pushMessage as jest.Mock).mockReturnValueOnce(JSON.stringify(emptyFullProof));
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Firefox);
-      mockGenerateRLNProof.mockResolvedValue(emptyFullProof);
-
-      const service = InjectorService.getInstance();
-
-      const result = await service.generateRLNProof(defaultProofRequest, defaultMetadata);
-
-      expect(result).toStrictEqual(emptyFullProof);
-      expect(pushMessage).toBeCalledTimes(0);
-    });
-
-    test("should throw error if generate proof is failed", async () => {
-      const error = new Error("error");
-      (pushMessage as jest.Mock).mockReturnValueOnce(JSON.stringify(emptyFullProof));
-      (getBrowserPlatform as jest.Mock).mockReturnValueOnce(BrowserPlatform.Firefox);
-      mockGenerateRLNProof.mockRejectedValue(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, defaultMetadata)).rejects.toThrowError(
-        `CryptKeeper: Error in generating RLN proof on Firefox error`,
-      );
-    });
-
-    test("should throw error there is no circuit and zkey files", async () => {
-      mockGetConnectedIdentity.mockResolvedValue({
-        serialize: () => mockSerializedIdentity,
-        genIdentityCommitment: () => "mockIdentityCommitment",
-      });
-      (browser.runtime.getURL as jest.Mock).mockReturnValue(undefined);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateRLNProof(defaultProofRequest, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: Must set RLN circuitFilePath and zkeyFilePath",
-      );
-    });
-  });
-
-  describe("Join Group", () => {
-    const defaultArgs: IJoinGroupMemberArgs = {
-      groupId: "90694543209366256629502773954857",
-      apiKey: "api-key",
-    };
-
-    beforeEach(() => {
-      (browser.runtime.getURL as jest.Mock).mockImplementation((path: string) => path);
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    test("should request group joining properly", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.joinGroup(defaultArgs, defaultMetadata)).resolves.toBeUndefined();
-      await expect(service.joinGroup({ groupId: defaultArgs.groupId }, defaultMetadata)).resolves.toBeUndefined();
-    });
-
-    test("should reject request group joining if urlOrigin isn't approved", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.joinGroup(defaultArgs, { urlOrigin: "new-urlOrigin" })).rejects.toThrow(
-        "CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.",
-      );
-      await expect(service.joinGroup({ groupId: defaultArgs.groupId }, { urlOrigin: "new-urlOrigin" })).rejects.toThrow(
-        "CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.",
-      );
-    });
-
-    test("should reject connect request from the approve connection page properly", async () => {
-      const error = new Error("User rejected your request.");
-      mockJoinGroupRequest.mockRejectedValue(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.joinGroup(defaultArgs, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: joining a group via Bandada service User rejected your request.",
-      );
-      await expect(service.joinGroup({ groupId: defaultArgs.groupId }, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: joining a group via Bandada service User rejected your request.",
-      );
-    });
-  });
-
-  describe("Generate Group Merkle Proof", () => {
-    const defaultArgs: IGenerateGroupMerkleProofArgs = {
-      groupId: "90694543209366256629502773954857",
-    };
-
-    beforeEach(() => {
-      (browser.runtime.getURL as jest.Mock).mockImplementation((path: string) => path);
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    test("should request group joining properly", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateGroupMerkleProof(defaultArgs, defaultMetadata)).resolves.toBeUndefined();
-      await expect(
-        service.generateGroupMerkleProof({ groupId: defaultArgs.groupId }, defaultMetadata),
-      ).resolves.toBeUndefined();
-    });
-
-    test("should reject request group joining if urlOrigin isn't approved", async () => {
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateGroupMerkleProof(defaultArgs, { urlOrigin: "new-urlOrigin" })).rejects.toThrow(
-        "CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.",
-      );
-      await expect(
-        service.generateGroupMerkleProof({ groupId: defaultArgs.groupId }, { urlOrigin: "new-urlOrigin" }),
-      ).rejects.toThrow("CryptKeeper: new-urlOrigin is not approved, please call 'connect()' request first.");
-    });
-
-    test("should reject connect request from the approve connection page properly", async () => {
-      const error = new Error("User rejected your request.");
-      mockGenerateGroupProofRequest.mockRejectedValue(error);
-
-      const service = InjectorService.getInstance();
-
-      await expect(service.generateGroupMerkleProof(defaultArgs, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: generate Merkle Proof via Bandada service User rejected your request.",
-      );
-      await expect(service.generateGroupMerkleProof({ groupId: defaultArgs.groupId }, defaultMetadata)).rejects.toThrow(
-        "CryptKeeper: generate Merkle Proof via Bandada service User rejected your request.",
       );
     });
   });
