--- packages/app/src/background/services/injector/types.ts
@@ -1,6 +1,0 @@
-import type { IRLNProofRequest, ISemaphoreProofRequest } from "@cryptkeeperzk/types";
-
-export interface IProofRequest<P = IRLNProofRequest | ISemaphoreProofRequest> {
-  identity: string;
-  payload: P;
-}
