+++ packages/app/src/background/services/protocol/index.ts
@@ -1,0 +1,3 @@
+import ProtocolService from "./ProtocolService";
+
+export default ProtocolService;
