+++ packages/types/src/approval/index.ts
@@ -1,5 +1,4 @@
 export interface IConnectionApprovalData {
-  checkedUrlOrigin: string;
   isApproved: boolean;
   canSkipApprove: boolean;
 }
